Rename file "operator.hpp" to "srfi-1.hpp" (resolve #73)
