+++ include/meevax/system/cursor.hpp
@@ -3,8 +3,8 @@
 
 #include <iterator>
 
-#include <meevax/system/operator.hpp>
 #include <meevax/system/pair.hpp>
+#include <meevax/system/srfi-1.hpp>
 
 namespace meevax::system
 {
