+++ include/scheme/base.hpp
@@ -6,8 +6,8 @@
 #include <meevax/system/boolean.hpp>
 #include <meevax/system/cursor.hpp>
 #include <meevax/system/number.hpp>
-#include <meevax/system/operator.hpp>
 #include <meevax/system/procedure.hpp>
+#include <meevax/system/srfi-1.hpp>
 
 extern "C"
 {
